Very Rough Encounter PoC for https://github.com/OpenPF2/PF2Core/issues/7
This is just a proof of concept that demonstrates:
1. Detecting enemies within 1 kilometer of the player (right now we just
   have a dummy enemy that does nothing but stand there).
2. Switching into encounter mode when the enemy is detected OR the a
   character triggers a "trap"/trigger box.
3. Queuing up character actions until their "turn" (right now this only
   works for the first player in the match and it just triggers whatever
   is queued up every six seconds).

More work remains in the framework to add:
- Initiative rolling in the MoPRS.
- Spending action points/applying costs.
- Triggering turn changes.

Also, the gun fire action is right now implemented via input bound to a
gameplay tag that triggers an ability granted to the character at
startup. This is only for debugging; the framework needs to add slotted
abilities next so that we can bind input to slots and then have slots
trigger whatever abilities they are bound to.