[#8] [#23] Ensure Activatable Abilities are Loaded on All Clients
It appears that `PossessedBy()` is called on listen servers but not
clients, while `OnRep_Controller()` is called on clients other than the
listen server. So... we need to attempt to load abilities for input in
multiple places. #FML

There may be a better injection point for this logic, but for now this
appears to work. I do not believe we want to do this when we're setting
up input because there's no guarantee that the character's abilities
have loaded on the server by that point. Perhaps down the road I'll need
to implement a pub-sub model where the bindings component listens to the
ASC directly to find out when abilities are granted, so we can load them
immediately instead of waiting for a call to load them en masse.