[#21] [#22] [#23] WIP: Encounter UI and Character Control Overhaul
This commit changes the Player State class to a new custom player state
Blueprint that uses the new OpenPF2 base class to implement the
PF2-specific Player State interface to provide the controllable
characters.

In addition, this refactors most of the UI to work with the new pawn
control logic that should soon allow us to display multiple controllable
characters to the player, and allow the player to toggle between them.

This does not yet work in Encounters, but should work very soon.