[#22] WIP Show Green Frame around Selected Character
This is currently driven off of which character wires up its command
bindings component to input, but that only works on clients and is not
sufficient for the server. We instead need to drive it off of what's getting
passed in to the `SetSelectedCharacter` method of the containing In-game UI UMG
widget, since that's guaranteed to get called on UI setup even if the input
wiring has already happened prior to the UI being initialized.