[#22] Rework Portrait Selection to Work through `SetSelectedCharacter`
Now, instead of character portraits listening for input to be connected
or disconnected for a character, we toggle a portrait as selected or
deselected through the containing `WBP_InGameUI` widget. The widget now
maintains a map of characters to portraits.

This design is necessary because the containing widget has to be
notified when there is a change in selected character, as it needs to
clear the command queue display and rebind event listeners to the newly
selected character.