[#22] Repopulate Command Queue Disp. when Switch Characters / Use Macros
Now, when switching characters, the command queue is cleared and
repopulated properly for whichever character is active.

This also converts several protected functions in the InGameUI blueprint
into being macros instead, as they are not referenced outside the
blueprint. (This is consistent with how macros are used elsewhere in the
project. I may decide to revisit this decision later if I find that this
is bloating memory usage.)