[#22] Replicate Bindings from Server + Move Ability Activation There
It doesn't seem like there's a way to activate abilities from bindings
properly unless they're invoked on the server. The problem is that at
the time that abilities are being activated on a character, the local
ASC might not have replicated its `ActivatableAbilities` and so the
client's view of the abilities will be inaccurate. This wasn't a
problem for the initial character that gets possessed, but characters
that the player possesses later were ending up with a blank ASC to
start.

This lead to the following error:
```
LogAbilitySystem: Warning: Can't activate LocalOnly or LocalPredicted
ability %s when not local!
```

The approach in this commit is not ideal; we should not have to go all
the way to the server for these actions to get invoked. It may be better
to go through native input handling in GAS.