[#21] Define Hexagon and Square Sprites for Movement
The core rulebook indicates that movement is designed around a square
grid system, but it appears that GMs sometimes prefer a hex grid to
simplify distance calculations.

According to the following post, hex grids are a "nightmare" for maps
that have man-made structures:
https://www.reddit.com/r/Pathfinder2e/comments/u5c0jg/why_didnt_pf2_make_the_jump_to_hexagons_instead/

For now, I'm defining sprites for either but might drop the hex grid
later.