[#21] First Working Ability to Hover to Highlight Movement Grid Cells
Now, a player can move the encounter cursor with the mouse over cells in
the movement grid to mark it as hovered. Each cell now turns green on
hover, and gray when not hovered.

In addition, the movement grid construction script is now more efficient
and cleaner. Rather than cloning a sample grid cell, there is now a
separate blueprint for grid cells that handles interfacing with the
encounter cursor and the grid.