[#21] Make Grid Cell Materials Dynamic and Adjust Hit Trace Tolerance
This makes the active and inactive materials of the grid as well as the
min and max trace tolerance for the cursor configurable on a per
instance basis.

In addition, the movement grid is now just below the character's feet
and the default trace distance is now from 2.5 meters above the
encounter cursor to 1 meter below it, to ensure that the player can
still select a location while on angled terrain. In a future version of
the project, I may be able to make the movement grid tesselate based on
the normal of the ground below it, but for now this movement grid will
work for our needs as a sample.