[#21] Introduce a "Virtual Cursor" for the Gamepad and Bind Left Stick
By default, UE only has support for projecting the mouse cursor into the
world. This adds a "virtual cursor" for the left analog stick of the
gamepad, and sources from that virtual cursor when input from a gamepad
is detected (we toggle back to the mouse cursor if mouse input is
detected).

To support this, input bindings on characters are set to no longer
consume input so that we can properly detect mouse input (e.g., for the
"Attack 1" input).