[#21] Clean-up Camera Pawn + Make Client Tilt Zoom Affect Transitions
Camera transitions should now all happen on the client side rather than
on the server. This not only frees-up server resources (since additional
clients do not force the server to plot transitions for multiple
cameras), but also improves UX. Now a player using a client can apply a
tilt-zoom during a transition and the transition will gracefully adapt
rather than ignoring the tilt-zoom until it "snaps" into the desired
location and rotation at the end of the transition.