[#21] Eliminate Transition Camera in Bird's Eye Camera; Simplify Interf.
"Set View Target with Blend" does what I was doing with the transition
camera and more -- it provides a nicer transition with more options. The
only snag with using it is that the PC can't automatically manage the
view target when switching pawns or the blend won't work; the view
target will automatically cut to the camera in the Bird's Eye pawn as
soon as it becomes possessed.

By making this change, the three methods of `BPI_PawnWithCamera` become
just a single method, `Activate Camera with Blend` in the new
`BPI_Camera` interface.