[#21] Limit Camera from Roaming Too Far Away + Fix Bug with NPC Enemy Scanning
This updates to the latest OpenPF2 to add new settings to the Bird's Eye Camera
Pawn ("Camera Pawn") that enables it to stop accepting movement input if it's
straying too far away from the character that the player is currently
controlling.

This also fixes a BP error that occurred whenever an NPC not aligned with a
party was scanned during the enemy detection tick. We now check if the owner
tracking component is present in the detected enemy before asking it what party
owns the enemy.