[#21] Fix Client-side Enc. Transitions with Multicast Auto. Cam. Control
View target transitions worked properly on clients until `2132aac`, when
the "transition camera" was removed. That new approach implemented a
client RPC that would disable automatic camera management before the
floating pawn was possessed, possess it, and then invoke the
"Set View Target with Blend" method on the client-side copy of the
player controller to transition the view from the old view to the
new view. It seems that this only worked on the listen server, though;
on clients, automatic camera management was still enabled during
possession, causing the switch between cameras to happe without a smooth
transition.

To fix this, the call to disable automatic camera management is now
Multicast. I am not 100% sure why this needs to be multicast, but with
it so, both the client and server transition into the floating camera
pawn view smoothly.