[#8] Separate Binding of "Default Action" from "Attack 1"
Getting UX the way that the player expects nearly always involves
something strange/non-standard from the developer perspective. Here, we
were binding "Face Button Bottom" (FBB), the left mouse button (LMB),
and the "1" key to "Attack 1" and then redirecting that to movement if
the character was in an encounter with a movement target. But, this
isn't the way that a player would expect things to work -- the player
would except that the "1" key invokes the primary attack, not movement,
while the LMB and FBB would invoke either the primary attack or movement
based on the context because that's the "default action".

Now, "1" is bound to the primary attack, while the LMB and FBB are bound
to a new "Default Action" input action that is filtered by an AEF to
determine if it should be for attack or movement.