[#8] WIP Changing Bindings in the Ability Dialog
This has temporarily modified the dialog not to disappear upon releasing
the SHIFT key/left trigger. With that disabled, the player can select an
ability and then press an input action and the binding will change. The
character still has to be toggled back and forth for the binding to take
effect, though.

Reset does not yet work.