[#8] [UE-78453] Implement Better ASC Owner Refresh during Encounters
In a multiplayer game, when controlling characters from a client other
than the listen server, the character the player was controlling before
entering an encounter would only appear to execute abilities and play
montages when viewed by other clients and the server, but not when
viewed by the client themself.

The problem relates to how characters are "unpossessed" at the start of
encounters, to enable the player to control the floating camera. When
a character gets unpossessed, both the server and the client change the
owner of the character pawn and then refresh the "Ability Owner"
information in the character's ASC. The server immediately knows that
the player's controller is no longer controlling the character but the
player's client does not receive this notification until the next frame,
so when it refreshes the ASC ability owner it does so with the wrong
controller. This forces the ASC to believe it is authoritative for all
the montages and abilities of the character when it is not.

The workaround is to override the `SetPawn()` method and provide
a timer callback on the character being unpossessed to update its ASC
during the next frame after controller ownership has replicated.

For more details on the bug, see:
https://github.com/OpenPF2/UE-78453-repro