[#53] Convert Dice Library to Non-Pure and Clarify Function Naming
Although dice rolls are just calculation with no side effects on other
system components, they are by definition random rather than
deterministic. This being the case, we can't have them be blueprint pure
because the engine may call a pure function at any time while evaluating
a graph and will even call them multiple times, one call per graph
reference. To avoid confusion for game designers and prevent the
disastrous side effect of the same dice roll returning a different
number through the same blueprint, it's much better to make this a
non-pure function so that Blueprint is forced to save the result of a
roll to a temporary local variable that remains the same value for the
rest of the graph.

While I'm here, it also makes sense to pick better display names for
these library functions so that it's clear from a cursory glance at the
graph that these are dice rolls. Previously, I was assuming that the
category and library name would be sufficient, but you don't have that
context for a graph node that's already been placed in a blueprint.