[#53] Re-factor Condition Icons out to Separate Widget + Add Level Text
These icons can be bound to any icon texture and any condition tag, now.
This makes it much easier to place these condition icons anywhere in a
character portrait, and increases cohesion for the functionality of
reacting to a condition level change (previously all this logic was in
the character portrait).